Specify the region when uploading bundles to s3
My previous understanding of the --region flag was wrong.  bundles are
uploaded to us-east-1 by default, regardless of the current region. One
fix is to upgrade and use the --region flag, but this can also be done
in older versions with a combination of --url and --location.

Since newer versions of ec2-ami-tools introduce many ubuntu-related
complications, fixing this with bundle-upload flags is preferable for now