Try new base-ami for ap-southeast-2
It keeps failing with the message:
Error registering AMI: Invalid value 'ami-67512a5d' for AMI. Can not set
kernel on an HVM AMI. (InvalidParameterValue)

No idea what the cause is, we may have to just disable the 14.04 hvm
builder in ap-southeast-2. First try an updated base ami, though, to see
if we stumbled on a bad base-ami build.