Remove upgrading section from the readme
Some of the advice was out of date (package names for cassandra 0.8 and
1.0?), some was wrong (installing any version of DSE other than the
latest fails unless you specify all the depedencies with versions as
well).

But most importantly, ComboAMI is aimed at test-clusters. It provides no
help with long-term the issues folks face in managing a long-lived
production clusters, and basically just delegates back to the
package-management system. So... in order to upgrade a ComboAMI cluster,
you upgrade as if you had installed from packages. Nothing to document
that isn't better handled elsewhere.